+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # developing
 - [httplib: fix unstable unit test which use the httplib.org](https://github.com/beego/beego/pull/5232)
 - [remove adapter package](https://github.com/beego/beego/pull/5239)
+- [feat: add write-delete cache mode](https://github.com/beego/beego/pull/5242)
 
 ## ORM refactoring
 - [introducing internal/models pkg](https://github.com/beego/beego/pull/5238)